requirements.txt
PyYAML~=6.0.1
pandas~=2.2.2
Authlib~=1.3.0
requests~=2.28.2
SQLAlchemy~=2.0.31
httpx~=0.23.3
python-dateutil~=2.8.2
schwab-py==1.2.0
seaborn
streamlit
boto3